Add transporters spoiler tab for dread Fixes wrong elevators for echoes Fixes #5097
