+++ randovania/games/dread/game_data.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from randovania.games import game
+from randovania.games.dread.gui.dread_teleporter_details_tab import DreadTeleporterDetailsTab
 from randovania.games.dread.layout.dread_configuration import DreadConfiguration
 from randovania.games.dread.layout.dread_cosmetic_patches import DreadCosmeticPatches
 from randovania.games.dread.layout.preset_describer import DreadPresetDescriber
@@ -20,7 +21,7 @@ def _gui() -> game.GameGui:
         cosmetic_dialog=gui.DreadCosmeticPatchesDialog,
         export_dialog=gui.DreadGameExportDialog,
         progressive_item_gui_tuples=progressive_items.tuples(),
-        spoiler_visualizer=(gui.DreadHintDetailsTab,),
+        spoiler_visualizer=(gui.DreadHintDetailsTab, DreadTeleporterDetailsTab),
         game_tab=gui.DreadGameTabWidget,
     )
