+++ randovania/games/dread/gui/dread_teleporter_details_tab.py
@@ -1,0 +1,23 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from randovania.games.dread.layout.dread_configuration import DreadConfiguration
+from randovania.gui.game_details.teleporter_details_tab import TeleporterDetailsTab
+
+if TYPE_CHECKING:
+    from logging.config import BaseConfigurator
+
+    from randovania.game_description.game_patches import GamePatches
+    from randovania.interface_common.players_configuration import PlayersConfiguration
+
+
+class DreadTeleporterDetailsTab(TeleporterDetailsTab):
+    @classmethod
+    def should_appear_for(cls, configuration: BaseConfigurator, all_patches: dict[int, GamePatches],
+                          players: PlayersConfiguration) -> bool:
+        assert isinstance(configuration, DreadConfiguration)
+        return not configuration.teleporters.is_vanilla
+
+    def tab_title(self) -> str:
+        return "Transporters"
