+++ randovania/gui/game_details/teleporter_details_tab.py
@@ -2,9 +2,6 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from randovania.games.game import RandovaniaGame
-from randovania.games.prime1.layout.prime_configuration import PrimeConfiguration
-from randovania.games.prime2.layout.echoes_configuration import EchoesConfiguration
 from randovania.gui.game_details.base_connection_details_tab import BaseConnectionDetailsTab
 from randovania.patching.prime import elevators
 
@@ -16,16 +13,10 @@ if TYPE_CHECKING:
 
 
 class TeleporterDetailsTab(BaseConnectionDetailsTab):
-    def tab_title(self) -> str:
-        if self.game_enum == RandovaniaGame.METROID_PRIME_CORRUPTION:
-            return "Teleporters"
-        return "Elevators"
-
     @classmethod
     def should_appear_for(cls, configuration: BaseConfiguration, all_patches: dict[int, GamePatches],
                           players: PlayersConfiguration) -> bool:
-        assert isinstance(configuration, PrimeConfiguration | EchoesConfiguration)
-        return not configuration.teleporters.is_vanilla
+        raise NotImplementedError
 
     def _fill_per_region_connections(self,
                                      per_region: dict[str, dict[str, str]],
@@ -33,6 +24,9 @@ class TeleporterDetailsTab(BaseConnectionDetailsTab):
                                      patches: GamePatches,
                                      ):
         for source, destination_loc in patches.all_dock_connections():
+            # Fix for portal rando showing up in echoes tab
+            if source.dock_type not in patches.game.dock_weakness_database.all_teleporter_dock_types:
+                continue
             source_region = region_list.region_by_area_location(source.identifier.area_identifier)
             source_name = elevators.get_elevator_or_area_name(self.game_enum, region_list,
                                                               source.identifier.area_identifier, True)
